Switch to Layout inspector on slide addition
