+++ editor/src/components/inspector/index.tsx
@@ -4,11 +4,15 @@ import { Tab, Tablist } from 'evergreen-ui';
 import { FormatInspector } from './format-inspector';
 import { DocumentInspector } from './document-inspector';
 import { InspectorTab, TabValuesList } from '../../types/inspector-tabs';
-import { useSwitchToFormatInspectorOnElementSelected } from '../../hooks/use-switch-to-inspector';
+import {
+  useSwitchToFormatInspectorOnElementSelected,
+  useSwitchToLayoutInspectorOnSlideAdded
+} from '../../hooks/inspector-hooks';
 
 export const Inspector = () => {
   const [activeTab, setActiveTab] = useState(InspectorTab.Document);
   useSwitchToFormatInspectorOnElementSelected({ setActiveTab });
+  useSwitchToLayoutInspectorOnSlideAdded({ setActiveTab });
 
   return (
     <InspectorContainer>
